miner: clean up BidBlock review fixes
